requirements3.txt
vosk
pyaudio # requires: brew install portaudio
language_tool_python
nltk
numpy
theano
pandas
openai
openai-whisper
asyncio
PySimpleGUI
tenacity
pyperclip
SpeechRecognition
torch